Restyle side navbar
Adds some CSS classes to the generated HTML output of the sidebar.
Allows for more control over styling in the future. The current styling
is more of an experiment and a WIP; doesn't fit in with the rest of
the page yet. This should be very easy to adjust and adapt, though.