Remove paddings from code blocks
The paddings of <code>-blocks introduced some inconsistent word spacings
in text, which is now resolved. This also means that, e.g., function signatures
are no longer indented/padded. This could be changed by adding custom CSS rules
for these cases explicitly, but the current layout should be reasonable, and
very similar to the layout of currydoc-4.