Add custom CSS style to export overview
The export overview's HTML structure has been improved,
and we can now freely adjust the CSS styling of the export entities.
Now, all entities are styled as cards, not just a simple list of
nav links. Furthermore, the overall robustness of the html document
was improved (resizing below a certain threshold hides the side navbar
properly and expands the main content, no font resizing, etc.).

Next, it would be nice to adjust the maximum content expansion
(e.g., using container-xl instead of container-fluid). For the moment,
this is hidden behind the Bootstrap4 curry library, though.