Fix indentation tracking in header
Before, when reaching a line with only whitespaces, the indetation of
the line before was discarded and the next line's indentation was picked
as a new baseline indentation. Because of the special nature of code block
formatting (indented by 4 whitespaces and separated by empty lines), blocks
were not generated properly.

Changing the tracking fixed this issue. Because a haddock-style documentation
comment starts with `-- |` and the ` -` was dropped, the indentation of sub-
sequent lines was incorrect and codeblocks following the first line of the
comment were indented improperly. Now, we replace all occurrences of `-- |`
by `--  `, resulting in consistent indentation based on the visual padding.